Move reconcileManifests() to tempostack_create_or_update.go
