+++ controllers/tempo/tempostack_controller_test.go
@@ -713,7 +713,7 @@ func TestReconcileManifestsValidateModes(t *testing.T) {
 			require.NoError(t, err)
 			reconciler := TempoStackReconciler{Client: k8sClient, Scheme: testScheme}
 			req := ctrl.Request{NamespacedName: nsn}
-			err = reconciler.reconcileManifests(context.Background(), logr.Discard(), req, *tempo)
+			err = reconciler.createOrUpdate(context.Background(), logr.Discard(), req, *tempo)
 			tc.validate(t, err)
 		})
 	}
